+++ app/src/main/java/com/example/androidacademy2/NewsListActivity.java
@@ -3,13 +3,21 @@ package com.example.androidacademy2;
 
 import android.content.Intent;
 import android.content.res.Configuration;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.view.View;
 import android.widget.HorizontalScrollView;
+import android.widget.ProgressBar;
+import android.widget.TextView;
 import android.widget.Toast;
 
+import java.util.List;
+import java.util.concurrent.ExecutionException;
+
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.DividerItemDecoration;
 import androidx.recyclerview.widget.LinearLayoutManager;
@@ -18,6 +26,10 @@ import androidx.recyclerview.widget.StaggeredGridLayoutManager;
 
 
 public class NewsListActivity extends AppCompatActivity {
+    @Nullable
+    private AsyncTask<Long, Void, List<NewsItem>> asyncTask;
+
+    public List<NewsItem> news;
 
     private final NewsRecyclerAdapter.OnItemClickListener clickListener = news ->
     {
@@ -35,24 +47,47 @@ public class NewsListActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_news_list);
+    }
 
-       // NewsItem news;
+    @Override
+    protected void onStart() {
+        super.onStart();
 
+        RecyclerView recyclerView = findViewById(R.id.recycler_news);
 
-        if (getResources().getConfiguration().orientation == Configuration.ORIENTATION_PORTRAIT) {
-            RecyclerView recyclerView = findViewById(R.id.recycler_news);
-            recyclerView.setAdapter(new NewsRecyclerAdapter(this, DataUtils.generateNews(), clickListener));
-            recyclerView.setLayoutManager(new LinearLayoutManager(this));
+        asyncTask = new LoadNews(this);
+        asyncTask.execute(1000L);
+        try {
+            news = asyncTask.get();
+            recyclerView.setVisibility(View.VISIBLE);
 
-            DividerItemDecoration dividerItemDecoration = new DividerItemDecoration(recyclerView.getContext(), 1);
-            recyclerView.addItemDecoration(dividerItemDecoration);
-        } else if (getResources().getConfiguration().orientation == Configuration.ORIENTATION_LANDSCAPE) {
-            RecyclerView recyclerView = findViewById(R.id.recycler_news);
-            recyclerView.setAdapter(new NewsRecyclerAdapter(this, DataUtils.generateNews(), clickListener));
-            recyclerView.setLayoutManager(new StaggeredGridLayoutManager(2, StaggeredGridLayoutManager.VERTICAL));
+            if (getResources().getConfiguration().orientation == Configuration.ORIENTATION_PORTRAIT) {
+                recyclerView.setAdapter(new NewsRecyclerAdapter(this, news, clickListener));
+                recyclerView.setLayoutManager(new LinearLayoutManager(this));
 
-            DividerItemDecoration dividerItemDecoration = new DividerItemDecoration(recyclerView.getContext(), 1);
-            recyclerView.addItemDecoration(dividerItemDecoration);
+                DividerItemDecoration dividerItemDecoration = new DividerItemDecoration(recyclerView.getContext(), 1);
+                recyclerView.addItemDecoration(dividerItemDecoration);
+            } else if (getResources().getConfiguration().orientation == Configuration.ORIENTATION_LANDSCAPE) {
+                recyclerView.setAdapter(new NewsRecyclerAdapter(this, news, clickListener));
+                recyclerView.setLayoutManager(new StaggeredGridLayoutManager(2, StaggeredGridLayoutManager.VERTICAL));
+
+                DividerItemDecoration dividerItemDecoration = new DividerItemDecoration(recyclerView.getContext(), 1);
+                recyclerView.addItemDecoration(dividerItemDecoration);
+            }
+        } catch (InterruptedException e) {
+            e.printStackTrace();
+        } catch (ExecutionException e) {
+            e.printStackTrace();
+        }
+
+    }
+
+    @Override
+    protected void onStop() {
+        super.onStop();
+        if (asyncTask != null) {
+            asyncTask.cancel(true);
+            asyncTask = null;
         }
     }
