+++ app/src/main/res/values/strings.xml
@@ -16,5 +16,5 @@
     <string name="menu_about">About</string>
     <string name="menu_news">News</string>
     <string name="image_news">Image news</string>
-    <string name="title_load">Load news</string>
+    <string name="load_complete">Download complete</string>
 </resources>
